Added routing to navbar along with a few page components
