+++ src/components/Home.js
@@ -1,0 +1,7 @@
+function Home() {
+    return (
+        <h1>Home</h1>
+    )
+}
+
+export default Home;
