+++ src/components/MyTrips.js
@@ -1,0 +1,7 @@
+function MyTrips() {
+    return (
+        <h1>My Trips</h1>
+    )
+}
+
+export default MyTrips;
