+++ src/components/Settings.js
@@ -1,0 +1,7 @@
+function Settings() {
+    return (
+        <h1>Settings</h1>
+    )
+}
+
+export default Settings;
